Stop assuming the journal knows what timestep the experiment consumes
This change defines the roles a little bit better. The journal has a history of
bit states, and it supplies this history to the VizModel whenever it does a
query. The VizModel is free to use this information. It uses yield, so the data
is only fetched when it's needed. Though a lazy sequence would be less fragile.

This unblocks potential future experiments where, for example, a higher layer
could connect to active bits from multiple previous timesteps.