Marshalling feature parity: channels
Before this change, the python server could receive channels but it couldn't
send them.

This change doesn't include a BigValueMarshal, so the client falls back to
sending big values every time. So this change is pretty bad for performance,
since it gets rid of the previous caching mechanism (case-by-case tokens).